// demo-api/src/schema/resolvers-types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]: Maybe<T[SubKey]> };
export type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>;
export type RequireFields<T, K extends keyof T> = {
  [X in Exclude<keyof T, K>]?: T[X];
} &
  { [P in K]-?: NonNullable<T[P]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

export enum Platform {
  Ios = "IOS",
  Android = "ANDROID",
}

export enum UpdateMechanism {
  Native = "NATIVE",
  Ota = "OTA",
}

export type Query = {
  __typename?: "Query";
  _?: Maybe<Scalars["Boolean"]>;
  controller?: Maybe<Controller>;
  controllers: Array<Controller>;
  features: FeatureMap;
  location?: Maybe<Location>;
  locations: Array<Location>;
  manufacturer: Manufacturer;
  me?: Maybe<User>;
  requestRating: Scalars["Boolean"];
  requestSurveyFeedback: Scalars["Boolean"];
  scheduleEvent?: Maybe<ScheduleEvent>;
  scheduleEvents: Array<ScheduleEvent>;
  temperaturePreset?: Maybe<TemperaturePreset>;
  temperaturePresets?: Maybe<Array<TemperaturePreset>>;
  updateRequired?: Maybe<UpdateMechanism>;
};

export type QueryControllerArgs = {
  id: Scalars["ID"];
};

export type QueryLocationArgs = {
  id: Scalars["ID"];
};

export type QueryRequestRatingArgs = {
  input?: Maybe<RequestRatingInput>;
};

export type QueryRequestSurveyFeedbackArgs = {
  input: RequestSurveyFeedbackInput;
};

export type QueryScheduleEventArgs = {
  id: Scalars["ID"];
};

export type QueryTemperaturePresetArgs = {
  id: Scalars["ID"];
};

export type QueryUpdateRequiredArgs = {
  input: UpdateRequiredInput;
};
// ...
export type RequestRatingInput = {
  build: Scalars["String"];
  installedAt: Scalars["String"];
  lastDisplayedAt?: Maybe<Scalars["String"]>;
  platform: Platform;
  version: Scalars["String"];
};

export type RequestSurveyFeedbackInput = {
  build: Scalars["String"];
  installedAt: Scalars["String"];
  lastDisplayedAt?: Maybe<Scalars["String"]>;
  lastResponseAt?: Maybe<Scalars["String"]>;
  platform: Platform;
  version: Scalars["String"];
};

export type UpdateRequiredInput = {
  platform: Platform;
  version: Scalars["String"];
  build: Scalars["String"];
};

export enum AppActiveTrackingFeature {
  Interval_60 = "INTERVAL_60",
}

export type FeatureMap = {
  __typename?: "FeatureMap";
  _?: Maybe<Scalars["Boolean"]>;
  accountSharing?: Maybe<AccountSharingFeature>;
  appActiveTracking?: Maybe<Array<AppActiveTrackingFeature>>;
  away?: Maybe<Array<AwayFeature>>;
  changeDefaultHoldLengthController?: Maybe<Array<DefaultHoldLengthFeature>>;
  changeDefaultHoldLengthLocation?: Maybe<Array<DefaultHoldLengthFeature>>;
  changeTemperatureUnit?: Maybe<ChangeTemperatureUnitFeature>;
  connect?: Maybe<ConnectFeature>;
  faultLogsController?: Maybe<FaultLogsFeature>;
  faultLogsLocation?: Maybe<FaultLogsFeature>;
  notifications?: Maybe<Array<NotificationsFeature>>;
  rename?: Maybe<Array<RenameFeature>>;
  schedule?: Maybe<ScheduleFeature>;
  signIn: SignInFeature;
  temperaturePresets?: Maybe<Array<TemperaturePresetsFeature>>;
};
// ...
export enum AccountSharingFeature {
  QrCode = "QR_CODE",
}

export enum SignInFeature {
  Token = "TOKEN",
}
// ...
export enum AwayFeature {
  AwayController = "AWAY_CONTROLLER",
  AwayLocation = "AWAY_LOCATION",
}

export type Away = {
  __typename?: "Away";
  active: Scalars["Boolean"];
  setpoint: Setpoint;
};

export type Controller = {
  __typename?: "Controller";
  activeHold?: Maybe<HoldLength>;
  away?: Maybe<Away>;
  call?: Maybe<Call>;
  defaultHoldLength?: Maybe<HoldLength>;
  fan?: Maybe<Fan>;
  faultActive?: Maybe<Scalars["Boolean"]>;
  faultLogs?: Maybe<Array<FaultLog>>;
  humidityAmbient?: Maybe<Scalars["Float"]>;
  humidityNotification?: Maybe<HumidityNotification>;
  id: Scalars["ID"];
  location: Location;
  mode: Mode;
  modes: Array<Mode>;
  name: Scalars["String"];
  schedule?: Maybe<Schedule>;
  setpoint: Setpoint;
  setpointRange: SetpointRange;
  temperatureAmbient: Scalars["Float"];
  temperatureNotification?: Maybe<TemperatureNotification>;
  zoning: Scalars["Boolean"];
};

export type Location = {
  __typename?: "Location";
  away?: Maybe<Away>;
  awayActive: Scalars["Boolean"];
  connectionStatus: ConnectionStatus;
  controller?: Maybe<Controller>;
  controllers: Array<Controller>;
  defaultHoldLength?: Maybe<HoldLength>;
  faultActive?: Maybe<Scalars["Boolean"]>;
  faultLogs?: Maybe<Array<FaultLog>>;
  faultNotification?: Maybe<BasicNotification>;
  humidityNotification?: Maybe<HumidityNotification>;
  id: Scalars["ID"];
  lat?: Maybe<Scalars["Float"]>;
  lng?: Maybe<Scalars["Float"]>;
  name: Scalars["String"];
  temperatureNotification?: Maybe<TemperatureNotification>;
  temperatureOutdoor?: Maybe<Scalars["Float"]>;
  temperaturePreset?: Maybe<TemperaturePreset>;
  temperaturePresets?: Maybe<Array<TemperaturePreset>>;
  temperatureUnit: TemperatureUnit;
  zoning: Scalars["Boolean"];
};
// ...
export enum Call {
  Heat = "HEAT",
  Cool = "COOL",
}

export type FanRunning = {
  running: Scalars["Boolean"];
};

export type SpeedNameFan = FanRunning & {
  __typename?: "SpeedNameFan";
  activeSpeedName?: Maybe<Scalars["String"]>;
  running: Scalars["Boolean"];
};

export type PercentageFan = FanRunning & {
  __typename?: "PercentageFan";
  activeSpeedPercent: Scalars["Float"];
  running: Scalars["Boolean"];
};

export type Fan = PercentageFan | SpeedNameFan;

export enum EffectiveMode {
  Cool = "COOL",
  Heat = "HEAT",
  Heatcool = "HEATCOOL",
  Off = "OFF",
}

export type Mode = {
  __typename?: "Mode";
  effectiveMode: EffectiveMode;
  name: Scalars["String"];
  placement: Placement;
  transitionFrom: Array<Mode>;
  transitionTo: Array<Mode>;
};
// ...
export enum ScheduleFeature {
  ScheduleTemperature = "SCHEDULE_TEMPERATURE",
  ScheduleTemperatureFan = "SCHEDULE_TEMPERATURE_FAN",
}

export enum Day {
  Sun = "SUN",
  Mon = "MON",
  Tue = "TUE",
  Wed = "WED",
  Thu = "THU",
  Fri = "FRI",
  Sat = "SAT",
}

export type ScheduleTime = {
  __typename?: "ScheduleTime";
  day: Day;
  hour: Scalars["Int"];
  minute: Scalars["Int"];
};

export type ScheduleTimeInput = {
  day: Day;
  hour: Scalars["Int"];
  minute: Scalars["Int"];
};

export type ScheduleEvent = {
  __typename?: "ScheduleEvent";
  id: Scalars["ID"];
  day: Day;
  removable: Scalars["Boolean"];
  start: ScheduleTime;
  end: ScheduleTime;
  nextEvent: ScheduleEvent;
  prevEvent: ScheduleEvent;
  temperaturePreset: TemperaturePreset;
};

export type ScheduleDay = {
  __typename?: "ScheduleDay";
  day: Day;
  events: Array<ScheduleEvent>;
  full: Scalars["Boolean"];
};

export type Schedule = {
  __typename?: "Schedule";
  days: Array<ScheduleDay>;
  monday: ScheduleDay;
  tuesday: ScheduleDay;
  wednesday: ScheduleDay;
  thursday: ScheduleDay;
  friday: ScheduleDay;
  saturday: ScheduleDay;
  sunday: ScheduleDay;
  maxEvents: Scalars["Int"];
  minEvents: Scalars["Int"];
  minEventInterval: Scalars["Int"];
};

export type HoldLengthIndefinite = {
  __typename?: "HoldLengthIndefinite";
  _?: Maybe<Scalars["Boolean"]>;
};

export type HoldLengthNextEvent = {
  __typename?: "HoldLengthNextEvent";
  _?: Maybe<Scalars["Boolean"]>;
};

export type HoldLengthHours = {
  __typename?: "HoldLengthHours";
  hours: Scalars["Int"];
};

export type HoldLengthDate = {
  __typename?: "HoldLengthDate";
  date: Scalars["String"];
};

export type HoldLength =
  | HoldLengthIndefinite
  | HoldLengthNextEvent
  | HoldLengthHours
  | HoldLengthDate;
// ...
export type Setpoint = SingleSetpoint | DualSetpoint;

export type SetpointRange = Range & {
  __typename?: "SetpointRange";
  min: Scalars["Float"];
  max: Scalars["Float"];
};

export type ChangeSetpointSuccess = {
  __typename?: "ChangeSetpointSuccess";
  controller: Controller;
};

export type AwayActive = Error & {
  __typename?: "AwayActive";
  message?: Maybe<Scalars["String"]>;
};

export type ChangeSetpointResult =
  | ChangeSetpointSuccess
  | AwayActive
  | NotSupported
  | NotFound;

export enum DefaultHoldLengthFeature {
  Indefinite = "INDEFINITE",
  NextEvent = "NEXT_EVENT",
  Hours_12 = "HOURS_12",
  Hours_24 = "HOURS_24",
  Date = "DATE",
}
// ...
export enum FaultLogsFeature {
  LogWithLabel = "LOG_WITH_LABEL",
  LogWithLabelAndDescription = "LOG_WITH_LABEL_AND_DESCRIPTION",
}

export type Log = {
  date: Scalars["String"];
};

export type FaultLogLabel = Log & {
  __typename?: "FaultLogLabel";
  date: Scalars["String"];
  label: Scalars["String"];
};

export type FaultLogLabelAndDescription = Log & {
  __typename?: "FaultLogLabelAndDescription";
  date: Scalars["String"];
  label: Scalars["String"];
  description: Scalars["String"];
};

export type FaultLog = FaultLogLabel | FaultLogLabelAndDescription;

export enum ConnectionStatus {
  Online = "ONLINE",
  Offline = "OFFLINE",
}
// ...
export enum ConnectFeature {
  AylaDisplay = "AYLA_DISPLAY",
}
// ...
export enum TemperatureUnit {
  F = "F",
  C = "C",
}

export enum ChangeTemperatureUnitFeature {
  AppOnly = "APP_ONLY",
}

export type ChangeTemperatureUnitInput = {
  id: Scalars["ID"];
  temperatureUnit: TemperatureUnit;
};

export type ChangeTemperatureUnitSuccess = {
  __typename?: "ChangeTemperatureUnitSuccess";
  location: Location;
};

export type ChangeTemperatureUnitResult =
  | ChangeTemperatureUnitSuccess
  | NotFound
  | NotSupported;

export type Manufacturer = {
  __typename?: "Manufacturer";
  support?: Maybe<Contact>;
};

export type Contact = {
  __typename?: "Contact";
  name: Scalars["String"];
  email?: Maybe<Scalars["String"]>;
  phone?: Maybe<Scalars["String"]>;
  website?: Maybe<Scalars["String"]>;
};

export enum NotificationsFeature {
  ControllerTemperature = "CONTROLLER_TEMPERATURE",
  ControllerHumidity = "CONTROLLER_HUMIDITY",
  LocationTemperature = "LOCATION_TEMPERATURE",
  LocationHumidity = "LOCATION_HUMIDITY",
  LocationFaults = "LOCATION_FAULTS",
}

export type Notification = {
  enabled: Scalars["Boolean"];
};

export type HumidityNotification = Notification &
  DualRangeValue & {
    __typename?: "HumidityNotification";
    enabled: Scalars["Boolean"];
    lower: PercentageRangeValue;
    upper: PercentageRangeValue;
    minInterval: Scalars["Float"];
  };

export type TemperatureNotification = Notification &
  DualRangeValue & {
    __typename?: "TemperatureNotification";
    enabled: Scalars["Boolean"];
    lower: SingleSetpoint;
    upper: SingleSetpoint;
    minInterval: Scalars["Float"];
  };

export type BasicNotification = Notification & {
  __typename?: "BasicNotification";
  enabled: Scalars["Boolean"];
};

export enum PushTokenStatus {
  Enabled = "ENABLED",
  Disabled = "DISABLED",
}

export type PushToken = {
  __typename?: "PushToken";
  id: Scalars["ID"];
  platform: Platform;
  status: PushTokenStatus;
  token: Scalars["String"];
};

export type User = {
  __typename?: "User";
  email: Scalars["String"];
  id: Scalars["String"];
  pushTokens?: Maybe<Array<PushToken>>;
};
// ...
export enum RenameFeature {
  Controller = "CONTROLLER",
  Location = "LOCATION",
}
// ...
export enum TemperaturePresetsFeature {
  BuiltIn = "BUILT_IN",
  Custom = "CUSTOM",
}

export enum Slot {
  Home = "HOME",
  Away = "AWAY",
  Sleep = "SLEEP",
  Custom = "CUSTOM",
}

export type TemperaturePreset = {
  __typename?: "TemperaturePreset";
  id: Scalars["ID"];
  name: Scalars["String"];
  slot?: Maybe<Slot>;
  setpoint: Setpoint;
  fanMode?: Maybe<FanMode>;
  removable: Scalars["Boolean"];
};
// ...
export type WithIndex<TObject> = TObject & Record<string, any>;
export type ResolversObject<TObject> = WithIndex<TObject>;

export type ResolverTypeWrapper<T> = Promise<T> | T;

export type LegacyStitchingResolver<TResult, TParent, TContext, TArgs> = {
  fragment: string;
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};

export type NewStitchingResolver<TResult, TParent, TContext, TArgs> = {
  selectionSet: string;
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export type StitchingResolver<TResult, TParent, TContext, TArgs> =
  | LegacyStitchingResolver<TResult, TParent, TContext, TArgs>
  | NewStitchingResolver<TResult, TParent, TContext, TArgs>;
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> =
  | ResolverFn<TResult, TParent, TContext, TArgs>
  | StitchingResolver<TResult, TParent, TContext, TArgs>;

export type ResolverFn<TResult, TParent, TContext, TArgs> = (
  parent: TParent,
  args: TArgs,
  context: TContext,
  info: GraphQLResolveInfo
) => Promise<TResult> | TResult;
// ...
export type ResolversTypes = ResolversObject<{
// ...
  HoldLength:
    | ResolversTypes["HoldLengthIndefinite"]
    | ResolversTypes["HoldLengthNextEvent"]
    | ResolversTypes["HoldLengthHours"]
    | ResolversTypes["HoldLengthDate"];
// ...
}>;

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = ResolversObject<{
// ...
  HoldLength:
    | ResolversParentTypes["HoldLengthIndefinite"]
    | ResolversParentTypes["HoldLengthNextEvent"]
    | ResolversParentTypes["HoldLengthHours"]
    | ResolversParentTypes["HoldLengthDate"];
// ...
}>;

export type QueryResolvers<
  ContextType = AppContext,
  ParentType extends ResolversParentTypes["Query"] = ResolversParentTypes["Query"]
> = ResolversObject<{
  _?: Resolver<Maybe<ResolversTypes["Boolean"]>, ParentType, ContextType>;
  controller?: Resolver<
    Maybe<ResolversTypes["Controller"]>,
    ParentType,
    ContextType,
    RequireFields<QueryControllerArgs, "id">
  >;
  controllers?: Resolver<
    Array<ResolversTypes["Controller"]>,
    ParentType,
    ContextType
  >;
  features?: Resolver<ResolversTypes["FeatureMap"], ParentType, ContextType>;
  location?: Resolver<
    Maybe<ResolversTypes["Location"]>,
    ParentType,
    ContextType,
    RequireFields<QueryLocationArgs, "id">
  >;
  locations?: Resolver<
    Array<ResolversTypes["Location"]>,
    ParentType,
    ContextType
  >;
  manufacturer?: Resolver<
    ResolversTypes["Manufacturer"],
    ParentType,
    ContextType
  >;
  me?: Resolver<Maybe<ResolversTypes["User"]>, ParentType, ContextType>;
  requestRating?: Resolver<
    ResolversTypes["Boolean"],
    ParentType,
    ContextType,
    RequireFields<QueryRequestRatingArgs, never>
  >;
  requestSurveyFeedback?: Resolver<
    ResolversTypes["Boolean"],
    ParentType,
    ContextType,
    RequireFields<QueryRequestSurveyFeedbackArgs, "input">
  >;
  scheduleEvent?: Resolver<
    Maybe<ResolversTypes["ScheduleEvent"]>,
    ParentType,
    ContextType,
    RequireFields<QueryScheduleEventArgs, "id">
  >;
  scheduleEvents?: Resolver<
    Array<ResolversTypes["ScheduleEvent"]>,
    ParentType,
    ContextType
  >;
  temperaturePreset?: Resolver<
    Maybe<ResolversTypes["TemperaturePreset"]>,
    ParentType,
    ContextType,
    RequireFields<QueryTemperaturePresetArgs, "id">
  >;
  temperaturePresets?: Resolver<
    Maybe<Array<ResolversTypes["TemperaturePreset"]>>,
    ParentType,
    ContextType
  >;
  updateRequired?: Resolver<
    Maybe<ResolversTypes["UpdateMechanism"]>,
    ParentType,
    ContextType,
    RequireFields<QueryUpdateRequiredArgs, "input">
  >;
}>;
// ...
export type HoldLengthResolvers<
  ContextType = AppContext,
  ParentType extends ResolversParentTypes["HoldLength"] = ResolversParentTypes["HoldLength"]
> = ResolversObject<{
  __resolveType: TypeResolveFn<
    | "HoldLengthIndefinite"
    | "HoldLengthNextEvent"
    | "HoldLengthHours"
    | "HoldLengthDate",
    ParentType,
    ContextType
  >;
}>;
// ...
export type Resolvers<ContextType = AppContext> = ResolversObject<{
// ...
  HoldLength?: HoldLengthResolvers<ContextType>;
// ...
}>;